Move Extensor back into example file.
